implement low-pass filtering in nuSQuIDS
This commit uses the low-pass filtering introduced in the latest PR to SQuIDS. This low-pass filtering can be applied while integrating the interaction picture states over time by calling
`Set_EvolLowPassCutoff` to choose a cut-off frequency of the filter. The frequency has units 1/length. This filter can speed up calculation of interaction picture states significantly in the presence of
eV-scale sterile neutrinos.

There are a few more additions to the Python interface of `nuSQUIDSLayers` that allow the filter to be applied also when projecting out probabilities from interpolated states.